requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.5
Cython==0.29.5
future==0.17.1
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pycparser==2.19
PyPDF2==1.26.0
pytesseract==0.2.6
python-telegram-bot==11.1.0
six==1.12.0
